Initial demo for viewing route data using mapbox
JS creates two html buttons, clicking on these adds the two available routes to the map. Or if the map already exists sets the bounds of the map to contain that route.
Also has functionality to add an empty route then add points to the route a coordinate at a time, could be used for live tracker